% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AggregateFolds}
\alias{AggregateFolds}
\title{Compute the mean for all the folds.
Operates on the output of MergeFolds().}
\usage{
AggregateFolds(cv.table)
}
\arguments{
\item{cv.table}{The output of MergeFolds().}

\item{n.params}{The number of parameters.}
}
\value{
A single data.table with all the folds aggregated.
}
\description{
Compute the mean for all the folds.
Operates on the output of MergeFolds().
}
