% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoparams.R
\docType{class}
\name{Autoparams-class}
\alias{Autoparams-class}
\title{Autoparams (S4 class)}
\description{
Autoparams (S4 class)
}
\section{Slots}{

\describe{
\item{\code{best}}{list.}

\item{\code{range}}{list.}

\item{\code{max_iter}}{numeric.}

\item{\code{improve_threshold}}{numeric.}

\item{\code{loss}}{function.}
}}

