% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoparams.R
\name{autoparams}
\alias{autoparams}
\title{Autoparams (constructor)}
\usage{
autoparams(start, range, loss = function(aggr) {     1 -
  aggr$pred_perf_R2 }, max_iter = 5, improve_threshold = 1)
}
\arguments{
\item{start}{Named list where each element corresponds to a parameter and stores the starting value.}

\item{range}{Names list where each element is a vector of two elements, min and max.}

\item{loss}{Function of aggr; its return value is minimised during the cross-validation.}

\item{max_iter}{Maximum number of successful iterations.}

\item{improve_threshold}{Percentage of the previous loss that must be achieved.}
}
\value{
An instance of the Autoparams class.
}
\description{
Autoparams (constructor)
}
