% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cites_year_plot.R
\name{cites_year_plot}
\alias{cites_year_plot}
\title{Plot number of citations per year (in Google Scholar).}
\usage{
cites_year_plot(user, fontsize = 18)
}
\arguments{
\item{user}{Google Scholar user ID.}

\item{fontsize}{Numeric. Font size for the plot.}
}
\value{
A plot.
}
\description{
Generate a ggplot2 figure of the number of total citations for a given researcher along time (years). Data from Google Scholar.
}
\examples{
\dontrun{
cites_year_plot(user="B7vSqZsAAAAJ")
}
}

