% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/citing_scholar.R
\name{citing_scholar}
\alias{citing_scholar}
\title{Citing articles in Google Scholar.}
\usage{
citing_scholar(doi)
}
\arguments{
\item{doi}{DOI of the paper.}
}
\value{
A URL as character.
}
\description{
Get URL of Scholar page with articles citing a given paper.
}
\examples{
\dontrun{
citing_scholar("http://dx.doi.org/10.1016/j.quascirev.2008.08.006")
}
}

