% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{compute_pearson}
\alias{compute_pearson}
\title{Title}
\usage{
compute_pearson(ytest, ypred, by = NULL, alternative = "g")
}
\arguments{
\item{ytest}{Vector of true values.}

\item{ypred}{Vector of predicted values.}

\item{by}{Factor by which to split the data.}

\item{alternative}{See cor.test().}
}
\value{
A "pearson" data.table.
}
\description{
Title
}
