% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\title{Cross-Validation}
\usage{
CV(d.train, model, params.grid = NULL, params.auto = NULL,
  n.folds = 5, folds.strata = NULL, model.strata = NULL,
  fit.best = FALSE, threads = 1, verbose = 1)
}
\arguments{
\item{d.train}{A dataset with both outcomes and predictors.}

\item{model}{A "model" function.}

\item{params.grid}{A list where each element is a vector containing all the values that are tried for that parameter.}

\item{params.auto}{An object of class "Autoparams" (see ?autoparams).}

\item{n.folds}{The number of folds.}

\item{folds.strata}{Factor according to which the folds are stratified.}

\item{model.strata}{Factor according to which the model is stratified.}

\item{fit.best}{Whether to fit the best model.}

\item{threads}{Number of threads.}

\item{verbose}{Whether to print what is being done.}
}
\value{
A list with table, aggr, and (if fit.best is TRUE) a fitted model.
}
\description{
Cross-Validation
}
