% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cvmse}
\alias{cvmse}
\title{MSE}
\usage{
cvmse(ytest, ypred)
}
\arguments{
\item{ytest}{A vector of the true response values.}

\item{ypred}{A vector of the predicted values.}

\item{p}{The number of predictors used by the model.}
}
\value{
The mean squared error of the model on the hold-out fold.
}
\description{
Mean squared error of the
}
