% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cvrsq}
\alias{cvrsq}
\title{\ifelse{html}{\out{R<sup>2</sup>}}{\eqn{R^2}}}
\usage{
cvrsq(ytest, ypred)
}
\arguments{
\item{ytest}{A vector of the true response values.}

\item{ypred}{A vector of the predicted values.}
}
\value{
The \ifelse{html}{\out{R<sup>2</sup>}}{\eqn{R^2}} of the model on the hold-out fold.
}
\description{
\ifelse{html}{\out{R<sup>2</sup>}}{\eqn{R^2}}
}
