% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CVtools-package.R
\docType{package}
\name{CVtools-package}
\alias{CVtools}
\alias{CVtools-package}
\title{CVtools package}
\description{
This package contains tools and functions to help prepare a modern, dynamic academic CV
(with citations, altmetrics, etc), plus CV analysis tools (e.g. citation or coauthorship networks).
It is particularly aimed for CVs maintained in Rmarkdown format.
}
\author{
Francisco Rodriguez-Sanchez
}
\keyword{package}

