% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvtools.R
\docType{package}
\name{cvtools}
\alias{cvtools}
\alias{cvtools-package}
\title{cvtools: Perform Cross-Validation and nested Cross-Validation.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

