% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cvvarfrac}
\alias{cvvarfrac}
\title{Fraction of Variability}
\usage{
cvvarfrac(ytest, ypred)
}
\arguments{
\item{ytest}{A vector of the true response values.}

\item{ypred}{A vector of the predicted values.}
}
\value{
The fraction of variability explained by the model in the hold-out fold.
}
\description{
Fraction of Variability
}
