% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_pubs_bibtex.R
\name{import_pubs_bibtex}
\alias{import_pubs_bibtex}
\title{Import publications from a BibTeX file and print them in Markdown format.}
\usage{
import_pubs_bibtex(file, ...)
}
\arguments{
\item{file}{Character. Path to bibtex file containing the references.}

\item{...}{Further parameters to pass to \code{\link{print.BibEntry}}, such as \code{max.names} (maximum number of names to display before substituting by 'et al.'), or \code{no.print.fields} to specify fields not to be printed (e.g. "doi", "url", "isbn", etc). Other options are fixed by default: \code{style} = 'text', \code{bib.style} = 'authoryear', \code{sorting} = 'ydnt', and \code{first.inits} = FALSE.}
}
\description{
Import publications from a BibTeX file and print them in Markdown format.
}
\author{
F. Rodriguez-Sanchez
}

