% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MergeFolds}
\alias{MergeFolds}
\title{Merge the folds to obtain a table with all the folds.
Operates on a list of "pearson" objects returned by Call_RF().}
\usage{
MergeFolds(cv.list)
}
\arguments{
\item{cv.list}{A list of "pearson" data.tables.}

\item{n.params}{The number of parameters.}
}
\value{
A single data.table with all the folds cbinded.
}
\description{
Merge the folds to obtain a table with all the folds.
Operates on a list of "pearson" objects returned by Call_RF().
}
