% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{NestedCV}
\alias{NestedCV}
\title{Nested Cross-Validation}
\usage{
NestedCV(d.train, model, params.grid = NULL, params.auto = NULL,
  n.outer.folds = 5, n.inner.folds = 10, folds.strata = NULL,
  model.strata = NULL, threads = 1, verbose = 1)
}
\arguments{
\item{d.train}{A dataset with both outcomes and predictors.}

\item{model}{A "model" function.}

\item{params.grid}{A list where each element is a vector containing all the values that are tried for that parameter.}

\item{params.auto}{An object of class "Autoparams" (see ?autoparams).}

\item{n.outer.folds}{Number of folds in the outer CV.}

\item{n.inner.folds}{Number of folds in the inner CV.}

\item{folds.strata}{Factor according to which the folds are stratified.}

\item{model.strata}{Factor according to which the model is stratified.}

\item{threads}{Number of threads.}

\item{verbose}{Whether to print what is being done.}
}
\value{
A list with inner CV, table, and aggr.
}
\description{
Nested Cross-Validation
}
