% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SummariseGenes}
\alias{SummariseGenes}
\title{Compute the mean for all the genes.
Operates on the output of AggregateFolds().}
\usage{
SummariseGenes(cv.aggr_table)
}
\arguments{
\item{cv.aggr_table}{The output of AggregateFolds().}

\item{n.params}{The number of parameters.}
}
\value{
A single data.table with all the rows summarised.
}
\description{
Compute the mean for all the genes.
Operates on the output of AggregateFolds().
}
