\name{cvTools-package}
\alias{cvTools-package}
\alias{cvTools}
\docType{package}
\title{Cross-validation tools for regression models}
\description{
Tools that allow developers to write functions
for cross-validation with minimal programming effort
and assist users with model selection.
}
\details{
\tabular{ll}{
Package: \tab cvTools\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2011-12-07\cr
Depends: \tab 
lattice,
robustbase\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Index:
\preformatted{
accessors               Access or set information on cross-validation
                        results
aggregate.cv            Aggregate cross-validation results
bwplot.cv               Box-and-whisker plots of cross-validation
                        results
cost                    Prediction loss
cvExamples              Cross-validation for linear models
cvFit                   Cross-validation for model evaluation
cvFolds                 Cross-validation folds
cvReshape               Reshape cross-validation results
cvSelect                Model selection based on cross-validation
cvTool                  Low-level function for cross-validation
cvTools-package         Cross-validation tools for regression models
cvTuning                Cross-validation for tuning parameter selection
densityplot.cv          Kernel density plots of cross-validation
                        results
dotplot.cvSelect        Dot plots of cross-validation results
plot.cv                 Plot cross-validation results
subset.cv               Subsetting cross-validation results
summary.cv              Summarize cross-validation results
xyplot.cvSelect         X-Y plots of cross-validation results
}
}
\author{
Andreas Alfons

Maintainer: Andreas Alfons
<andreas.alfons@econ.kuleuven.be>
}
\keyword{package}
